Split the file, unfmt
